# GPU accelerated backend for interpolation functions