--- a/sarracen/interpolate.py
+++ b/sarracen/interpolate.py
@@ -1,10 +1,13 @@
 """
 Contains several interpolation functions which produce interpolated 2D or 1D arrays of SPH data.
 """
+import math
+
 import numpy as np
-from numba import prange, njit
+from numba import prange, njit, cuda
 from scipy.spatial.transform import Rotation as R, Rotation
 
+from sarracen._atomic_operations import atomic_add
 from sarracen.kernels import BaseKernel
 
 
@@ -238,7 +241,7 @@ def _rotate_data(data, x, y, z, rotation, origin):
 
 def interpolate_2d(data: 'SarracenDataFrame', target: str, x: str = None, y: str = None, kernel: BaseKernel = None,
                    x_pixels: int = None, y_pixels: int = None,  x_min: float = None, x_max: float = None,
-                   y_min: float = None, y_max: float = None) -> np.ndarray:
+                   y_min: float = None, y_max: float = None, backend: str = 'cpu') -> np.ndarray:
     """ Interpolate particle data across two directional axes to a 2D grid of pixels.
 
     Interpolate the data within a SarracenDataFrame to a 2D grid, by interpolating the values
@@ -289,8 +292,8 @@ def interpolate_2d(data: 'SarracenDataFrame', target: str, x: str = None, y: str
     _check_dimension(data, 2)
 
     return _fast_2d(data[target].to_numpy(), 0, data[x].to_numpy(), data[y].to_numpy(), np.zeros(len(data)),
-                    data['m'].to_numpy(), data['rho'].to_numpy(), data['h'].to_numpy(), kernel.w, kernel.get_radius(),
-                    x_pixels, y_pixels, x_min, x_max, y_min, y_max, 2)
+                        data['m'].to_numpy(), data['rho'].to_numpy(), data['h'].to_numpy(), kernel.w,
+                        kernel.get_radius(), x_pixels, y_pixels, x_min, x_max, y_min, y_max, 2, backend)
 
 
 def interpolate_2d_cross(data: 'SarracenDataFrame',
@@ -302,7 +305,8 @@ def interpolate_2d_cross(data: 'SarracenDataFrame',
                          x1: float = None,
                          x2: float = None,
                          y1: float = None,
-                         y2: float = None) -> np.ndarray:
+                         y2: float = None,
+                         backend: str = 'cpu') -> np.ndarray:
     """ Interpolate particle data across two directional axes to a 1D cross-section line.
 
     Interpolate the data within a SarracenDataFrame to a 1D line, by interpolating the values
@@ -355,7 +359,7 @@ def interpolate_2d_cross(data: 'SarracenDataFrame',
 
     return _fast_2d_cross(data[target].to_numpy(), data[x].to_numpy(), data[y].to_numpy(), data['m'].to_numpy(),
                           data['rho'].to_numpy(), data['h'].to_numpy(), kernel.w, kernel.get_radius(), pixels, x1,
-                          x2, y1, y2)
+                          x2, y1, y2, backend)
 
 
 def interpolate_3d(data: 'SarracenDataFrame',
@@ -371,7 +375,8 @@ def interpolate_3d(data: 'SarracenDataFrame',
                    x_min: float = None,
                    x_max: float = None,
                    y_min: float = None,
-                   y_max: float = None):
+                   y_max: float = None,
+                   backend: str = 'cpu'):
     """ Interpolate 3D particle data to a 2D grid of pixels.
 
     Interpolates three-dimensional particle data in a SarracenDataFrame. The data
@@ -435,9 +440,11 @@ def interpolate_3d(data: 'SarracenDataFrame',
     kernel = kernel if kernel is not None else data.kernel
     _check_dimension(data, 3)
 
+    weight_function = kernel.get_column_kernel_func(integral_samples)
+
     return _fast_2d(data[target].to_numpy(), 0, x_data, y_data, np.zeros(len(data)), data['m'].to_numpy(),
-                    data['rho'].to_numpy(), data['h'].to_numpy(), kernel.get_column_kernel_func(integral_samples),
-                    kernel.get_radius(), x_pixels, y_pixels, x_min, x_max, y_min, y_max, 2)
+                        data['rho'].to_numpy(), data['h'].to_numpy(), weight_function,
+                        kernel.get_radius(), x_pixels, y_pixels, x_min, x_max, y_min, y_max, 2, backend)
 
 
 def interpolate_3d_cross(data: 'SarracenDataFrame',
@@ -454,7 +461,8 @@ def interpolate_3d_cross(data: 'SarracenDataFrame',
                          x_min: float = None,
                          x_max: float = None,
                          y_min: float = None,
-                         y_max: float = None):
+                         y_max: float = None,
+                         backend: str = 'cpu'):
     """ Interpolate 3D particle data to a 2D grid, using a 3D cross-section.
 
     Interpolates particle data in a SarracenDataFrame across three directional axes to a 2D
@@ -527,16 +535,36 @@ def interpolate_3d_cross(data: 'SarracenDataFrame',
 
     x_data, y_data, z_data = _rotate_data(data, x, y, z, rotation, origin)
 
-    return _fast_2d(data[target].to_numpy(), z_slice, x_data, y_data, z_data,
-                    data['m'].to_numpy(), data['rho'].to_numpy(), data['h'].to_numpy(), kernel.w,
-                    kernel.get_radius(), x_pixels, y_pixels, x_min, x_max, y_min, y_max, 3)
+    return _fast_2d(data[target].to_numpy(), z_slice, x_data, y_data, z_data, data['m'].to_numpy(),
+                        data['rho'].to_numpy(), data['h'].to_numpy(), kernel.w, kernel.get_radius(), x_pixels, y_pixels,
+                        x_min, x_max, y_min, y_max, 3, backend)
 
 
-# Underlying numba-compiled code for interpolation to a 2D grid. Used in interpolation of 2D data,
+def _fast_2d(target, z_slice, x_data, y_data, z_data, mass_data, rho_data, h_data, weight_function, kernel_radius,
+             x_pixels, y_pixels, x_min, x_max, y_min, y_max, n_dims, backend):
+    if backend == 'cpu':
+        return _fast_2d_cpu(target, z_slice, x_data, y_data, z_data, mass_data, rho_data, h_data, weight_function,
+                            kernel_radius, x_pixels, y_pixels, x_min, x_max, y_min, y_max, n_dims)
+    elif backend == 'gpu':
+        return _fast_2d_gpu(target, z_slice, x_data, y_data, z_data, mass_data, rho_data, h_data, weight_function,
+                            kernel_radius, x_pixels, y_pixels, x_min, x_max, y_min, y_max, n_dims)
+
+
+def _fast_2d_cross(target, x_data, y_data, mass_data, rho_data, h_data, weight_function, kernel_radius, pixels, x1, x2,
+                   y1, y2, backend):
+    if backend == 'cpu':
+        return _fast_2d_cross_cpu(target, x_data, y_data, mass_data, rho_data, h_data, weight_function,
+                                  kernel_radius, pixels, x1, x2, y1, y2)
+    elif backend == 'gpu':
+        return _fast_2d_cross_gpu(target, x_data, y_data, mass_data, rho_data, h_data, weight_function,
+                                  kernel_radius, pixels, x1, x2, y1, y2)
+
+
+# Underlying CPU numba-compiled code for interpolation to a 2D grid. Used in interpolation of 2D data,
 # and column integration / cross-sections of 3D data.
 @njit(parallel=True, fastmath=True)
-def _fast_2d(target, z_slice, x_data, y_data, z_data, mass_data, rho_data, h_data, weight_function, kernel_radius,
-             x_pixels, y_pixels, x_min, x_max, y_min, y_max, n_dims):
+def _fast_2d_cpu(target, z_slice, x_data, y_data, z_data, mass_data, rho_data, h_data, weight_function, kernel_radius,
+                 x_pixels, y_pixels, x_min, x_max, y_min, y_max, n_dims):
     image = np.zeros((y_pixels, x_pixels))
     pixwidthx = (x_max - x_min) / x_pixels
     pixwidthy = (y_max - y_min) / y_pixels
@@ -580,18 +608,106 @@ def _fast_2d(target, z_slice, x_data, y_data, z_data, mass_data, rho_data, h_dat
 
         # calculate contributions at pixels i, j due to particle at x, y
         q2 = dx2i + dy2.reshape(len(dy2), 1)
-        wab = weight_function(np.sqrt(q2), n_dims)
 
-        # add contributions to image
-        image[jpixmin:jpixmax, ipixmin:ipixmax] += (wab * term[i])
+        for jpix in prange(jpixmax - jpixmin):
+            for ipix in prange(ipixmax - ipixmin):
+                if np.sqrt(q2[jpix][ipix]) > kernel_radius:
+                    continue
+                wab = weight_function(np.sqrt(q2[jpix][ipix]), n_dims)
+                atomic_add(image, (jpix + jpixmin, ipix + ipixmin), term[i] * wab)
 
     return image
 
 
+# Underlying numba-compiled code for interpolation to a 2D grid. Used in interpolation of 2D data,
+# and column integration / cross-sections of 3D data.
+# For the GPU, the numba code is compiled using a factory function approach. This results in less
+# overhead for the weight function.
+def _fast_2d_gpu(target, z_slice, x_data, y_data, z_data, mass_data, rho_data, h_data, weight_function, kernel_radius,
+                 x_pixels, y_pixels, x_min, x_max, y_min, y_max, n_dims):
+    threadsperblock = 32
+    blockspergrid = (target.size + (threadsperblock - 1)) // threadsperblock
+
+    # transfer relevant data to the GPU
+    d_target = cuda.to_device(target)
+    d_x = cuda.to_device(x_data)
+    d_y = cuda.to_device(y_data)
+    d_z = cuda.to_device(z_data)
+    d_m = cuda.to_device(mass_data)
+    d_rho = cuda.to_device(rho_data)
+    d_h = cuda.to_device(h_data)
+    d_image = cuda.to_device(np.zeros((y_pixels, x_pixels)))
+
+    @cuda.jit(fastmath=True)
+    def _2d_func(target, z_slice, x_data, y_data, z_data, mass_data, rho_data, h_data, kernel_radius,
+                     x_pixels, y_pixels, x_min, x_max, y_min, y_max, n_dims, image):
+        pixwidthx = (x_max - x_min) / x_pixels
+        pixwidthy = (y_max - y_min) / y_pixels
+
+        i = cuda.grid(1)
+        if i < len(target):
+            if not n_dims == 2:
+                dz = np.float64(z_slice) - z_data[i]
+            else:
+                dz = 0
+
+            term = (target[i] * mass_data[i] / (rho_data[i] * h_data[i] ** n_dims))
+
+            if abs(dz) >= kernel_radius * h_data[i]:
+                return
+
+            # determine maximum and minimum pixels that this particle contributes to
+            ipixmin = round((x_data[i] - kernel_radius * h_data[i] - x_min) / pixwidthx)
+            jpixmin = round((y_data[i] - kernel_radius * h_data[i] - y_min) / pixwidthy)
+            ipixmax = round((x_data[i] + kernel_radius * h_data[i] - x_min) / pixwidthx)
+            jpixmax = round((y_data[i] + kernel_radius * h_data[i] - y_min) / pixwidthy)
+
+            if ipixmax < 0 or ipixmin > x_pixels or jpixmax < 0 or jpixmin > y_pixels:
+                return
+            if ipixmin < 0:
+                ipixmin = 0
+            if ipixmax > x_pixels:
+                ipixmax = x_pixels
+            if jpixmin < 0:
+                jpixmin = 0
+            if jpixmax > y_pixels:
+                jpixmax = y_pixels
+
+            # calculate contributions to all nearby pixels
+            for jpix in range(jpixmax - jpixmin):
+                for ipix in range(ipixmax - ipixmin):
+                    # determine difference in the x-direction
+                    xpix = x_min + ((ipix + ipixmin) + 0.5) * pixwidthx
+                    dx = xpix - x_data[i]
+                    dx2 = dx * dx * (1 / (h_data[i] ** 2))
+
+                    # determine difference in the y-direction
+                    ypix = y_min + ((jpix + jpixmin) + 0.5) * pixwidthy
+                    dy = ypix - y_data[i]
+                    dy2 = dy * dy * (1 / (h_data[i] ** 2))
+
+                    dz2 = ((dz ** 2) * (1 / h_data[i] ** 2))
+
+                    # calculate contributions at pixels i, j due to particle at x, y
+                    q = math.sqrt(dx2 + dy2 + dz2)
+
+                    # add contribution to image
+                    if q < kernel_radius:
+                        # atomic add protects the summation against race conditions.
+                        wab = weight_function(q, n_dims)
+                        cuda.atomic.add(image, (jpix + jpixmin, ipix + ipixmin), term * wab)
+
+    _2d_func[blockspergrid, threadsperblock](d_target, z_slice, d_x, d_y, d_z, d_m, d_rho, d_h,
+                                                 kernel_radius, x_pixels, y_pixels, x_min,
+                                                 x_max, y_min, y_max, n_dims, d_image)
+
+    return d_image.copy_to_host()
+
+
 # Underlying numba-compiled code for 2D->1D cross-sections
 @njit(parallel=True, fastmath=True)
-def _fast_2d_cross(target, x_data, y_data, mass_data, rho_data, h_data, weight_function, kernel_radius, pixels, x1, x2,
-                   y1, y2):
+def _fast_2d_cross_cpu(target, x_data, y_data, mass_data, rho_data, h_data, weight_function, kernel_radius, pixels, x1,
+                       x2, y1, y2):
     # determine the slope of the cross-section line
     gradient = 0
     if not x2 - x1 == 0:
@@ -647,7 +763,87 @@ def _fast_2d_cross(target, x_data, y_data, mass_data, rho_data, h_data, weight_f
         q2 = (dx * dx + dy * dy) * (1 / (h_data[filter_det][i] * h_data[filter_det][i]))
         wab = weight_function(np.sqrt(q2), 2)
 
-        # add contributions to output total, transformed by minimum/maximum pixels
-        output[int(ipixmin[i]):int(ipixmax[i])] += (wab * term[filter_det][i])
+        # add contributions to output total
+        for ipix in prange(int(ipixmax[i]) - int(ipixmin[i])):
+            atomic_add(output, ipix + int(ipixmin[i]), term[i] * wab[ipix])
 
     return output
+
+
+# Underlying numba-compiled code for 2D->1D cross-sections
+def _fast_2d_cross_gpu(target, x_data, y_data, mass_data, rho_data, h_data, weight_function, kernel_radius, pixels, x1,
+                       x2, y1, y2):
+    threadsperblock = 32
+    blockspergrid = (target.size + (threadsperblock - 1)) // threadsperblock
+
+    # transfer relevant data to the GPU
+    d_target = cuda.to_device(target)
+    d_x = cuda.to_device(x_data)
+    d_y = cuda.to_device(y_data)
+    d_m = cuda.to_device(mass_data)
+    d_rho = cuda.to_device(rho_data)
+    d_h = cuda.to_device(h_data)
+    d_image = cuda.to_device(np.zeros(pixels))
+
+    # determine the slope of the cross-section line
+    gradient = 0
+    if not x2 - x1 == 0:
+        gradient = (y2 - y1) / (x2 - x1)
+    yint = y2 - gradient * x2
+
+    # determine the fraction of the line that one pixel represents
+    xlength = math.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
+    pixwidth = xlength / pixels
+    xpixwidth = (x2 - x1) / pixels
+    aa = 1 + gradient ** 2
+
+    @cuda.jit(fastmath=True)
+    def _2d_func(target, x_data, y_data, mass_data, rho_data, h_data, kernel_radius, pixels, x1, x2, y1, y2, image):
+        i = cuda.grid(1)
+        if i < target.size:
+            term = target[i] * mass_data[i] / (rho_data[i] * h_data[i] ** 2)
+
+            # the intersections between the line and a particle's 'smoothing circle' are
+            # found by solving a quadratic equation with the below values of a, b, and c.
+            # if the determinant is negative, the particle does not contribute to the
+            # cross-section, and can be removed.
+            bb = 2 * gradient * (yint - y_data[i]) - 2 * x_data[i]
+            cc = x_data[i] ** 2 + y_data[i] ** 2 - 2 * yint * y_data[i] + yint ** 2 - (kernel_radius * h_data[i]) ** 2
+            det = bb ** 2 - 4 * aa * cc
+
+            # create a filter for particles that do not contribute to the cross-section
+            if det < 0:
+                return
+
+            det = math.sqrt(det)
+
+            # the starting and ending x coordinates of the lines intersections with a particle's smoothing circle
+            xstart = min(max(x1, (-bb - det) / (2 * aa)), x2)
+            xend = min(max(y1, (-bb + det) / (2 * aa)), y2)
+
+            # the start and end distances which lie within a particle's smoothing circle.
+            rstart = math.sqrt((xstart - x1) ** 2 + ((gradient * xstart + yint) - y1) ** 2)
+            rend = math.sqrt((xend - x1) ** 2 + (((gradient * xend + yint) - y1) ** 2))
+
+            # the maximum and minimum pixels that each particle contributes to.
+            ipixmin = min(max(0, round(rstart / pixwidth)), pixels)
+            ipixmax = min(max(0, round(rend / pixwidth)), pixels)
+
+            # iterate through all affected pixels
+            for ipix in range(ipixmin, ipixmax):
+                # determine contributions to all affected pixels for this particle
+                xpix = x1 + (ipix + 0.5) * xpixwidth
+                ypix = gradient * xpix + yint
+                dy = ypix - y_data[i]
+                dx = xpix - x_data[i]
+
+                q2 = (dx * dx + dy * dy) * (1 / (h_data[i] * h_data[i]))
+                wab = weight_function(math.sqrt(q2), 2)
+
+                # add contributions to output total, transformed by minimum/maximum pixels
+                cuda.atomic.add(image, ipix, wab * term)
+
+    _2d_func[blockspergrid, threadsperblock](d_target, d_x, d_y, d_m, d_rho, d_h, kernel_radius, pixels, x1, x2, y1, y2,
+                                             d_image)
+
+    return d_image.copy_to_host()